moved utils to root package
