+++ Turnauswertung-py3/utils/txt.py
